Move architecture to dedicated folder and update codex
