+++ architecture/architecture.js
